Use the mjs extension for the modern bundle name
Newer versions of Node assume that js file extension is for CommonJS
files when `package.json` does not contain `type: "module"`. This
prevents importing the package on newer Node versions, since the
`index.modern.js` is assumed to be CJS whereas it has `export`
statements which end in a SyntaxError.

Let's use the mjs extension for the modern bundle as recommended in
microbundle's README [0].

[0]: https://github.com/developit/microbundle#specifying-filenames-in-packagejson